more unified DMA allocation - memory allocator is put into a separte independent module, snd-alloc-pages. it can be kept after other snd* modules unloaded, so that the big continuous pages can be held after reload. - use size_t for the size type.

+++ sound/pci/es1968.c
@@ -557,7 +557,7 @@ struct snd_es1968 {
 	/* buffer */
 	void *dma_buf;
 	dma_addr_t dma_buf_addr;
-	unsigned long dma_buf_size;
+	size_t dma_buf_size;
 
 	/* Resources... */
 	int irq;
